refactor: 키보드 hide 기능 구현 및 persistentContainer 재배치 #9
