+++ CloudNotes/CloudNotes/Model/CoreDataManager.swift
@@ -1,4 +1,5 @@
 import Foundation
+import CoreData
 
 class CoreDataManager {
     static let shared = CoreDataManager()
@@ -6,11 +7,36 @@ class CoreDataManager {
     weak var memoListViewController: MemoReloadable?
     weak var memoContentViewController: MemoReloadable?
     
+    private let persistentContainer: NSPersistentCloudKitContainer = {
+        let container = NSPersistentCloudKitContainer(name: "CloudNotes")
+        container.loadPersistentStores(completionHandler: { (_, error) in
+            if let error = error as NSError? {
+                fatalError("Unresolved error \(error), \(error.userInfo)")
+            }
+        })
+        
+        return container
+    }()
+    
+    private var context: NSManagedObjectContext {
+        return persistentContainer.viewContext
+    }
+
     private init() { }
+
+    func saveContextChange () {
+        if context.hasChanges {
+            do {
+                try context.save()
+            } catch {
+                let nserror = error as NSError
+                fatalError("Unresolved error \(nserror), \(nserror.userInfo)")
+            }
+        }
+    }
     
     func load() -> [Memo] {
         do {
-            let context = AppDelegate.persistentContainer.viewContext
             return try context.fetch(Memo.fetchRequest()).reversed()
         } catch {
             print(error)
@@ -20,7 +46,6 @@ class CoreDataManager {
     }
     
     func delete(data: Memo) {
-        let context = AppDelegate.persistentContainer.viewContext
         context.delete(data)
         do {
             try context.save()
@@ -35,7 +60,6 @@ class CoreDataManager {
         data.title = title
         data.body = body
         data.lastModified = Date().timeIntervalSince1970
-        let context = AppDelegate.persistentContainer.viewContext
         do {
             try context.save()
             memoListViewController?.reload()
@@ -45,7 +69,6 @@ class CoreDataManager {
     }
     
     func create() {
-        let context = AppDelegate.persistentContainer.viewContext
         let memo = Memo(context: context)
         memo.lastModified = Date().timeIntervalSince1970
         do {
